+++ parser/config.go
@@ -143,6 +143,8 @@ var SupportedFunctions = map[string]int{
 	"digitsum":   1,
 	"primorial": 1,
 	"choose":    2,
+	"omega":     1,
+	"bigomega":  1,
 	"sumofprimes": 1,
 	"lambertw":  1,
 	"zeta":      1,